SCAL-636 Tests refactoring.
