--- test/supervisor_script_helper.rb
@@ -1,2 +1,0 @@
-module SupervisorScriptHelper
-end
